Added functionality to FixMultiplication Class
